Create remove_duplicates_(sorted_array)
As the first element of the array will always be original , i start traversal along the array from "1" position till the end of the array.
in the traversal , i will check the current position element with the previous element . the counter "k" is to keep track of the unique elements and also act as a pointer which rewrites the array with only unique elements. In the end , k is returned as number of unique elements in the array.